+++ android/src/main/java/com/opentokreactnative/OTSessionManager.java
@@ -18,13 +18,17 @@ import com.facebook.react.bridge.ReactContext;
 import com.facebook.react.bridge.ReactContextBaseJavaModule;
 import com.facebook.react.bridge.ReactMethod;
 import com.facebook.react.bridge.UiThreadUtil;
+import com.facebook.react.bridge.WritableArray;
 import com.facebook.react.bridge.WritableMap;
 import com.facebook.react.modules.core.DeviceEventManagerModule;
 import com.facebook.react.bridge.ReadableMap;
 import com.facebook.react.bridge.ReadableArray;
+import com.facebook.react.bridge.Promise;
 
 import com.opentok.android.Session;
 import com.opentok.android.Connection;
+import com.opentok.android.MediaUtils;
+import com.opentok.android.MuteForcedInfo;
 import com.opentok.android.Publisher;
 import com.opentok.android.PublisherKit;
 import com.opentok.android.Stream;
@@ -47,13 +51,19 @@ public class OTSessionManager extends ReactContextBaseJavaModule
         implements Session.SessionListener,
         PublisherKit.PublisherListener,
         PublisherKit.AudioLevelListener,
+        PublisherKit.PublisherRtcStatsReportListener,
+        PublisherKit.AudioStatsListener,
+        PublisherKit.MuteListener,
+        PublisherKit.VideoStatsListener,
         SubscriberKit.SubscriberListener,
         Session.SignalListener,
         Session.ConnectionListener,
         Session.ReconnectionListener,
         Session.ArchiveListener,
+        Session.MuteListener,
         Session.StreamPropertiesListener,
         SubscriberKit.AudioLevelListener,
+        SubscriberKit.SubscriberRtcStatsReportListener,
         SubscriberKit.AudioStatsListener,
         SubscriberKit.VideoStatsListener,
         SubscriberKit.VideoListener,
@@ -119,6 +129,7 @@ public class OTSessionManager extends ReactContextBaseJavaModule
         mSession.setReconnectionListener(this);
         mSession.setArchiveListener(this);
         mSession.setStreamPropertiesListener(this);
+        mSession.setMuteListener(this);
         mSessions.put(sessionId, mSession);
         mAndroidOnTopMap.put(sessionId, androidOnTop);
         mAndroidZOrderMap.put(sessionId, androidZOrder);
@@ -153,6 +164,7 @@ public class OTSessionManager extends ReactContextBaseJavaModule
         Boolean publishAudio = properties.getBoolean("publishAudio");
         Boolean publishVideo = properties.getBoolean("publishVideo");
         String videoSource = properties.getString("videoSource");
+        Boolean scalableScreenshare = properties.getBoolean("scalableScreenshare");
         Publisher mPublisher = null;
         if (videoSource.equals("screen")) {
             View view = getCurrentActivity().getWindow().getDecorView().getRootView();
@@ -166,6 +178,7 @@ public class OTSessionManager extends ReactContextBaseJavaModule
                     .resolution(Publisher.CameraCaptureResolution.valueOf(resolution))
                     .frameRate(Publisher.CameraCaptureFrameRate.valueOf(frameRate))
                     .capturer(capturer)
+                    .scalableScreenshare(scalableScreenshare)
                     .build();
             mPublisher.setPublisherVideoType(PublisherKit.PublisherKitVideoType.PublisherKitVideoTypeScreen);
         } else {
@@ -186,9 +199,13 @@ public class OTSessionManager extends ReactContextBaseJavaModule
         }
         mPublisher.setPublisherListener(this);
         mPublisher.setAudioLevelListener(this);
+        mPublisher.setRtcStatsReportListener(this);
         mPublisher.setAudioFallbackEnabled(audioFallbackEnabled);
         mPublisher.setPublishVideo(publishVideo);
         mPublisher.setPublishAudio(publishAudio);
+        mPublisher.setAudioStatsListener(this);
+        mPublisher.setVideoStatsListener(this);
+        mPublisher.setMuteListener(this);
         ConcurrentHashMap<String, Publisher> mPublishers = sharedState.getPublishers();
         mPublishers.put(publisherId, mPublisher);
         callback.invoke();
@@ -227,6 +244,7 @@ public class OTSessionManager extends ReactContextBaseJavaModule
         mSubscriber.setAudioLevelListener(this);
         mSubscriber.setAudioStatsListener(this);
         mSubscriber.setVideoStatsListener(this);
+        mSubscriber.setRtcStatsReportListener(this);
         mSubscriber.setVideoListener(this);
         mSubscriber.setStreamListener(this);
         mSubscriber.setSubscribeToAudio(properties.getBoolean("subscribeToAudio"));
@@ -242,6 +260,9 @@ public class OTSessionManager extends ReactContextBaseJavaModule
                     preferredResolution.getInt("width"),
                     preferredResolution.getInt("height"));
             mSubscriber.setPreferredResolution(resolution);
+        }
+        if (properties.hasKey("audioVolume")) {
+            mSubscriber.setAudioVolume((float) properties.getDouble("audioVolume"));
         }
         mSubscribers.put(streamId, mSubscriber);
         if (mSession != null) {
@@ -291,6 +312,59 @@ public class OTSessionManager extends ReactContextBaseJavaModule
     }
 
     @ReactMethod
+    public void forceMuteAll(String sessionId, ReadableArray excludedStreamIds, Promise promise) {
+        ConcurrentHashMap<String, Session> mSessions = sharedState.getSessions();
+        Session mSession = mSessions.get(sessionId);
+        ConcurrentHashMap<String, Stream> streams = sharedState.getSubscriberStreams();
+        ArrayList<Stream> mExcludedStreams = new ArrayList<Stream>();
+        if (mSession == null) {
+            promise.reject("Session not found.");
+            return;
+        }
+        for (int i = 0; i < excludedStreamIds.size(); i++) {
+            String streamId = excludedStreamIds.getString(i);
+            Stream mStream = streams.get(streamId);
+            if (mStream == null) {
+                promise.reject("Stream not found.");
+                return;
+            }
+            mExcludedStreams.add(mStream);
+        }
+        mSession.forceMuteAll(mExcludedStreams);
+        promise.resolve(null);
+    }
+
+    @ReactMethod
+    public void forceMuteStream(String sessionId, String streamId, Promise promise) {
+        ConcurrentHashMap<String, Session> mSessions = sharedState.getSessions();
+        Session mSession = mSessions.get(sessionId);
+        ConcurrentHashMap<String, Stream> streams = sharedState.getSubscriberStreams();
+        if (mSession == null) {
+            promise.reject("Session not found.");
+            return;
+        }
+        Stream mStream = streams.get(streamId);
+        if (mStream == null) {
+            promise.reject("Stream not found.");
+            return;
+        }
+        mSession.forceMuteStream(mStream);
+        promise.resolve(null);
+    }
+
+    @ReactMethod
+    public void disableForceMute(String sessionId, Promise promise) {
+        ConcurrentHashMap<String, Session> mSessions = sharedState.getSessions();
+        Session mSession = mSessions.get(sessionId);
+        if (mSession == null) {
+            promise.reject("Session not found.");
+            return;
+        }
+        mSession.disableForceMute();
+        promise.resolve(true);
+    }
+
+    @ReactMethod
     public void publishAudio(String publisherId, Boolean publishAudio) {
 
         ConcurrentHashMap<String, Publisher> mPublishers = sharedState.getPublishers();
@@ -307,6 +381,16 @@ public class OTSessionManager extends ReactContextBaseJavaModule
         Publisher mPublisher = mPublishers.get(publisherId);
         if (mPublisher != null) {
             mPublisher.setPublishVideo(publishVideo);
+        }
+    }
+
+    @ReactMethod
+    public void getRtcStatsReport(String publisherId) {
+
+        ConcurrentHashMap<String, Publisher> mPublishers = sharedState.getPublishers();
+        Publisher mPublisher = mPublishers.get(publisherId);
+        if (mPublisher != null) {
+            mPublisher.getRtcStatsReport();
         }
     }
 
@@ -356,6 +440,34 @@ public class OTSessionManager extends ReactContextBaseJavaModule
         if (mSubscriber != null) {
             mSubscriber.setPreferredFrameRate(frameRate);
         }
+    }
+
+    @ReactMethod
+    public void setAudioVolume(String streamId, Float audioVolume) {
+
+        ConcurrentHashMap<String, Subscriber> mSubscribers = sharedState.getSubscribers();
+        Subscriber mSubscriber = mSubscribers.get(streamId);
+        if (mSubscriber != null) {
+            mSubscriber.setAudioVolume(audioVolume);
+        }
+    }
+
+    @ReactMethod
+    public void getSubscriberRtcStatsReport(String streamId) {
+
+        ConcurrentHashMap<String, Subscriber> mSubscribers = sharedState.getSubscribers();
+        Subscriber mSubscriber = mSubscribers.get(streamId);
+        if (mSubscriber != null) {
+            mSubscriber.getRtcStatsReport();
+        }
+    }
+
+    @ReactMethod
+    public void getSupportedCodecs(Promise promise) {
+
+        MediaUtils.SupportedCodecs mSupportedCodecs = MediaUtils.getSupportedCodecs(this.getReactApplicationContext());
+        WritableMap supportedCodecsMap = EventUtils.prepareMediaCodecsMap(mSupportedCodecs);
+        promise.resolve(supportedCodecsMap); 
     }
 
     @ReactMethod
@@ -493,6 +605,33 @@ public class OTSessionManager extends ReactContextBaseJavaModule
     }
 
     @ReactMethod
+    public void getSessionCapabilities(String sessionId, Callback callback) {
+        ConcurrentHashMap<String, Session> mSessions = sharedState.getSessions();
+        Session mSession = mSessions.get(sessionId);
+        WritableMap sessionCapabilitiesMap = Arguments.createMap();
+        if (mSession != null){
+            Session.Capabilities sessionCapabilities = mSession.getCapabilities();
+            sessionCapabilitiesMap.putBoolean("canForceMute", sessionCapabilities.canForceMute);
+            sessionCapabilitiesMap.putBoolean("canPublish", sessionCapabilities.canPublish);
+            // Bug in OT Android SDK. This should always be true, but it is set to false:
+            sessionCapabilitiesMap.putBoolean("canSubscribe", true);
+        }
+        callback.invoke(sessionCapabilitiesMap);
+    }
+
+    @ReactMethod
+    public void reportIssue(String sessionId, Callback callback) {
+        ConcurrentHashMap<String, Session> mSessions = sharedState.getSessions();
+        Session mSession = mSessions.get(sessionId);
+        if (mSession != null){
+          callback.invoke(mSession.reportIssue());
+        } else {
+          callback.invoke(null, "Error connecting to session. Could not find native session instance.");
+
+        }
+    }
+
+    @ReactMethod
     public void enableLogs(Boolean logLevel) {
         setLogLevel(logLevel);
     }
@@ -507,6 +646,15 @@ public class OTSessionManager extends ReactContextBaseJavaModule
             reactContext
                     .getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter.class)
                     .emit(eventName, eventData);
+        }
+    }
+
+    private void sendEventArray(ReactContext reactContext, String eventName, @Nullable WritableArray eventData) {
+
+        if (Utils.contains(jsEvents, eventName) || Utils.contains(componentEvents, eventName)) {
+            reactContext
+                .getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter.class)
+                .emit(eventName, eventData);
         }
     }
 
@@ -662,6 +810,17 @@ public class OTSessionManager extends ReactContextBaseJavaModule
         sendEventMap(this.getReactApplicationContext(), session.getSessionId() + ":" + sessionPreface + "onStreamDropped", streamInfo);
         printLogs("onStreamDropped: Stream Dropped: "+stream.getStreamId() +" in session: "+session.getSessionId());
     }
+    @Override
+    public void onMuteForced​(Session session, MuteForcedInfo info) {
+
+        WritableMap muteForcedInfo = Arguments.createMap();
+        String sessionId = session.getSessionId();
+        muteForcedInfo.putString("sessionId", sessionId);
+        Boolean active = info.getActive();
+        muteForcedInfo.putBoolean("active", active);
+        sendEventMap(this.getReactApplicationContext(), session.getSessionId() + ":" + sessionPreface + "onMuteForced​", muteForcedInfo);
+        printLogs("Mute forced -- active: " + active + " in session: " + sessionId);
+    }
 
     @Override
     public void onStreamCreated(PublisherKit publisherKit, Stream stream) {
@@ -718,6 +877,49 @@ public class OTSessionManager extends ReactContextBaseJavaModule
         if (publisherId.length() > 0) {
             String event = publisherId + ":" + publisherPreface + "onAudioLevelUpdated";
             sendEventWithString(this.getReactApplicationContext(), event, String.valueOf(audioLevel));
+        }
+    }
+
+    @Override
+    public void onRtcStatsReport(PublisherKit publisher, PublisherKit.PublisherRtcStats[] stats) {
+
+        String publisherId = Utils.getPublisherId(publisher);
+        if (publisherId.length() > 0) {
+            WritableArray rtcStatsReportArray = EventUtils.preparePublisherRtcStats(stats);
+            String event = publisherId + ":" + publisherPreface + "onRtcStatsReport";
+            sendEventArray(this.getReactApplicationContext(), event, rtcStatsReportArray);
+        }
+    }
+
+    @Override
+    public void onAudioStats(PublisherKit publisher, PublisherKit.PublisherAudioStats[] stats) {
+
+        String publisherId = Utils.getPublisherId(publisher);
+        if (publisherId.length() > 0) {
+            WritableArray publisherInfo = EventUtils.preparePublisherAudioStats(stats);
+            String event = publisherId + ":" + publisherPreface + "onAudioStats";
+            sendEventArray(this.getReactApplicationContext(), event, publisherInfo);
+        }
+    }
+
+    @Override
+    public void onVideoStats(PublisherKit publisher, PublisherKit.PublisherVideoStats[] stats) {
+
+        String publisherId = Utils.getPublisherId(publisher);
+        if (publisherId.length() > 0) {
+            WritableArray publisherInfo = EventUtils.preparePublisherVideoStats(stats);
+            String event = publisherId + ":" + publisherPreface +  "onVideoStats";
+            sendEventArray(this.getReactApplicationContext(), event, publisherInfo);
+        }
+    }
+
+    @Override
+    public void onMuteForced​(PublisherKit publisher) {
+
+        String publisherId = Utils.getPublisherId(publisher);
+        if (publisherId.length() > 0) {
+            String event = publisherId + ":" + publisherPreface + "onMuteForced";
+            sendEventMap(this.getReactApplicationContext(), event, null);
         }
     }
 
@@ -786,6 +988,22 @@ public class OTSessionManager extends ReactContextBaseJavaModule
         printLogs("onError: "+opentokError.getErrorDomain() + " : " +
                 opentokError.getErrorCode() +  " - "+opentokError.getMessage());
 
+    }
+
+    @Override
+    public void onRtcStatsReport(SubscriberKit subscriberKit, String stats) {
+
+        String streamId = Utils.getStreamIdBySubscriber(subscriberKit);
+        if (streamId.length() > 0) {
+            ConcurrentHashMap<String, Stream> streams = sharedState.getSubscriberStreams();
+            Stream mStream = streams.get(streamId);
+            WritableMap subscriberInfo = Arguments.createMap();
+            if (mStream != null) {
+                subscriberInfo.putMap("stream", EventUtils.prepareJSStreamMap(mStream, subscriberKit.getSession()));
+            }
+            subscriberInfo.putString("jsonArrayOfReports", stats);
+            sendEventMap(this.getReactApplicationContext(), subscriberPreface +  "onRtcStatsReport", subscriberInfo);
+        }
     }
 
     @Override