+++ src/OTSubscriber.js
@@ -53,6 +53,9 @@ export default class OTSubscriber extends Component {
         if (preferredFrameRate !== undefined) {
           OT.setPreferredFrameRate(streamId, sanitizeFrameRate(preferredFrameRate));
         }
+        if (audioVolume !== undefined) {
+          OT.setAudioVolume(streamId, audioVolume);
+        }
       });
       this.setState({ streamProperties });
     }
@@ -98,6 +101,9 @@ export default class OTSubscriber extends Component {
       }
     });
   }
+  getRtcStatsReport(streamId) {
+    OT.getSubscriberRtcStatsReport(streamId);
+  }
   render() {
     if (!this.props.children) {
       const containerStyle = this.props.containerStyle;
@@ -120,6 +126,7 @@ OTSubscriber.propTypes = {
   eventHandlers: PropTypes.object, // eslint-disable-line react/forbid-prop-types
   streamProperties: PropTypes.object, // eslint-disable-line react/forbid-prop-types
   containerStyle: PropTypes.object, // eslint-disable-line react/forbid-prop-types
+  getRtcStatsReport: PropTypes.object, // eslint-disable-line react/forbid-prop-types
   subscribeToSelf: PropTypes.bool
 };
 
@@ -128,7 +135,8 @@ OTSubscriber.defaultProps = {
   eventHandlers: {},
   streamProperties: {},
   containerStyle: {},
-  subscribeToSelf: false
+  subscribeToSelf: false,
+  getRtcStatsReport: {},
 };
 
 OTSubscriber.contextType = OTContext;